+++ src/code42cli/cmds/securitydata/main.py
@@ -9,6 +9,7 @@ from code42cli.cmds.securitydata.extraction import extract
 from code42cli.cmds.search_shared.cursor_store import FileEventCursorStore
 from code42cli.commands import Command, SubcommandLoader
 from code42cli.cmds.securitydata.savedsearch.commands import SavedSearchSubCommandLoader
+from code42cli.cmds.securitydata.savedsearch import savedsearch
 
 
 class SecurityDataSubcommandLoader(SubcommandLoader):
@@ -77,19 +78,25 @@ def clear_checkpoint(sdk, profile):
 def print_out(sdk, profile, args):
     """Activates 'print' command. It gets security events and prints them to stdout."""
     logger = logger_factory.get_logger_for_stdout(args.format)
-    extract(sdk, profile, logger, args)
+    query = sdk.securitydata.savedsearches.get_query(args.saved_search) \
+        if args.saved_search else None
+    extract(sdk, profile, logger, args, query)
 
 
 def write_to(sdk, profile, args):
     """Activates 'write-to' command. It gets security events and writes them to the given file."""
     logger = logger_factory.get_logger_for_file(args.output_file, args.format)
-    extract(sdk, profile, logger, args)
+    query = sdk.securitydata.savedsearches.get_query(args.saved_search) \
+        if args.saved_search else None
+    extract(sdk, profile, logger, args, query)
 
 
 def send_to(sdk, profile, args):
     """Activates 'send-to' command. It gets security events and logs them to the given server."""
     logger = logger_factory.get_logger_for_server(args.server, args.protocol, args.format)
-    extract(sdk, profile, logger, args)
+    query = sdk.securitydata.savedsearches.get_query(args.saved_search) \
+        if args.saved_search else None
+    extract(sdk, profile, logger, args, query)
 
 
 def _load_write_to_args(arg_collection):
@@ -112,6 +119,11 @@ def _load_send_to_args(arg_collection):
 
     arg_collection.extend(send_to_args)
     _load_search_args(arg_collection)
+
+
+def _load_saved_search_args(arg_collection):
+    saved_search = ArgConfig(u"--saved-search", help=u"saved search id.")
+    arg_collection.append(u"saved_search", saved_search)
 
 
 def _load_search_args(arg_collection):
@@ -177,4 +189,5 @@ def _load_search_args(arg_collection):
         ),
     }
     search_args = args.create_search_args(search_for=u"file events", filter_args=filter_args)
+    _load_saved_search_args(arg_collection)
     arg_collection.extend(search_args)